refactor: move a bit of code from core to cli package
