+++ .dockerignore
@@ -5,7 +5,6 @@
 .github
 .history
 bin/
-cli/
 dashboard/public/
 dashboard/src/
 dist/
