+++ cli/src/add-version-files.ts
