+++ cli/src/generate-docs.ts
