+++ core/package.json
