+++ lerna.json
@@ -2,7 +2,8 @@
   "packages": [
     "cli/",
     "core/",
-    "dashboard/"
+    "dashboard/",
+    "sdk/"
   ],
   "stream": true,
   "version": "independent"
