+++ sdk/.prettierrc.json
@@ -1,0 +1,1 @@
+../.prettierrc.json
