+++ sdk/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@garden-io/cli",
+  "name": "@garden-io/sdk",
   "version": "0.12.2",
-  "description": "Cloud native testing platform for testing and developing container applications on Kubernetes",
+  "description": "TypeScript plugin SDK for Garden",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/garden-io/garden.git"
@@ -18,50 +18,28 @@
   "files": [
     "build"
   ],
-  "bin": {
-    "garden": "bin/garden"
-  },
+  "main": "build/index.js",
   "dependencies": {
     "@garden-io/core": "0.12.2"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.12",
-    "@types/mocha": "^8.0.3",
-    "bluebird": "^3.7.2",
-    "chai": "^4.2.0",
-    "chalk": "^4.1.0",
-    "execa": "^4.0.3",
-    "lodash": "^4.17.20",
-    "minimist": "^1.2.5",
-    "mocha": "^8.1.1",
-    "mocha-logger": "^1.0.6",
-    "pkg": "4.4.9",
-    "prettier": "^2.1.0",
-    "split2": "^3.2.2",
+    "@types/node": "^12.12.29",
+    "prettier": "^1.19.1",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "tslint-no-unused": "^0.2.0-alpha.1",
     "tslint-plugin-prettier": "^2.3.0",
-    "typescript": "^3.9.6",
-    "username": "^5.1.0"
+    "typescript": "^3.9.6"
   },
   "scripts": {
-    "add-version-files": "node build/src/add-version-files.js",
-    "build": "npm run tsc && npm run add-version-files && npm run generate-docs",
+    "build": "npm run tsc",
     "check-package-lock": "git diff-index --quiet HEAD -- package-lock.json || (echo 'package-lock.json is dirty!' && exit 1)",
-    "clean": "shx rm -rf build dist",
+    "clean": "shx rm -rf build",
     "dev": "npm run watch",
     "fix-format": "prettier --write \"{src,test}/**/*.ts\"",
-    "generate-docs": "node ./build/src/generate-docs.js",
     "lint": "tslint -p .",
     "tsc": "tsc -p .",
     "watch": "npm run tsc -- -w"
-  },
-  "pkg": {
-    "scripts": [
-      "package.json",
-      "node_modules/@garden-io/core/build/src/**/*.js"
-    ]
   }
 }
