+++ sdk/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../tsconfig.base.json",
   "include": [
-    "src/*.ts",
+    "src/**/*.ts",
     "test/**/*.ts"
   ],
   "compilerOptions": {
